% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.BASGRA.R
\name{met2model.BASGRA}
\alias{met2model.BASGRA}
\title{Write BASGRA met files}
\usage{
met2model.BASGRA(in.path, in.prefix, outfolder, overwrite = FALSE,
  start_date, end_date, ...)
}
\arguments{
\item{in.path}{path on disk where CF file lives}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}

\item{start_date}{beginning of the weather data}

\item{end_date}{end of the weather data}
}
\value{
OK if everything was succesful.
}
\description{
Converts a met CF file to a model specific met file. The input
files are calld <in.path>/<in.prefix>.YYYY.cf
}
\author{
Istem Fer
}
