% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.BASGRA.R
\name{read_restart.BASGRA}
\alias{read_restart.BASGRA}
\title{Read restart function for SDA with BASGRA}
\usage{
read_restart.BASGRA(outdir, runid, stop.time, settings, var.names, params)
}
\value{
X.vec      vector of forecasts
}
\description{
Read Restart for BASGRA
}
\author{
Istem Fer
}
