% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.BASGRA.R
\name{write.config.BASGRA}
\alias{write.config.BASGRA}
\title{Write BASGRA configuration files}
\usage{
write.config.BASGRA(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}

\item{trait.samples}{vector of samples for a given trait}
}
\value{
configuration file for BASGRA for given run
}
\description{
Writes a BASGRA config file.
}
\details{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Istem Fer
}
