% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.BASGRA.R
\name{write_restart.BASGRA}
\alias{write_restart.BASGRA}
\title{write_restart.BASGRA}
\usage{
write_restart.BASGRA(
  outdir,
  runid,
  start.time,
  stop.time,
  settings,
  new.state,
  RENAME = TRUE,
  new.params = FALSE,
  inputs
)
}
\value{
TRUE if successful
}
\description{
Write restart files for BASGRA
}
\author{
Istem Fer
}
