% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_BASGRA.R
\name{run_BASGRA}
\alias{run_BASGRA}
\title{run BASGRA model}
\usage{
run_BASGRA(
  run_met,
  run_params,
  site_harvest,
  site_fertilize,
  start_date,
  end_date,
  outdir,
  sitelat,
  sitelon,
  co2_file = NULL,
  write_raw_output = FALSE
)
}
\arguments{
\item{run_met}{path to CF met}

\item{run_params}{parameter vector}

\item{site_harvest}{path to harvest file}

\item{site_fertilize}{path to fertilizer application file}

\item{start_date}{start time of the simulation}

\item{end_date}{end time of the simulation}

\item{outdir}{where to write BASGRA output}

\item{sitelat}{latitude of the site}

\item{sitelon}{longitude of the site}

\item{co2_file}{path to daily atmospheric CO2 concentration file, optional, defaults to 350 ppm when missing}

\item{write_raw_output}{write raw output in csv or not}
}
\description{
BASGRA wrapper function. Runs and writes model outputs in PEcAn standard.
}
\details{
BASGRA is written in fortran is run through R by wrapper functions written by Marcel Van Oijen.
This function makes use of those wrappers but gives control of datastream in and out of the model to PEcAn.
With this function we skip model2netcdf, we can also skip met2model but keeping it for now. 
write.config.BASGRA modifies args of this function through template.job
then job.sh runs calls this function to run the model
}
\author{
Istem Fer, Julius Vira
}
