% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.BASGRA.R
\name{write_restart.BASGRA}
\alias{write_restart.BASGRA}
\title{write_restart.SIPNET}
\usage{
write_restart.BASGRA(
  outdir,
  runid,
  start.time,
  stop.time,
  settings,
  new.state,
  RENAME = TRUE,
  new.params = FALSE,
  inputs
)
}
\arguments{
\item{outdir}{outout directory}

\item{runid}{run id}

\item{start.time}{Time of current assimilation step}

\item{stop.time}{Time of next assimilation step}

\item{settings}{pecan settings list}

\item{new.state}{Analysis state matrix returned by \code{sda.enkf}}

\item{RENAME}{flag to either rename output file or not}

\item{new.params}{optional, additionals params to pass write.configs that are deterministically related to the parameters updated by the analysis}

\item{inputs}{new input paths updated by the SDA workflow, will be passed to write.configs}
}
\value{
TRUE if successful
}
\description{
Write restart files for BASGRA
}
\author{
Istem Fer
}
